WIP. fixing stuff
