+++ Sample/Podfile
@@ -1,5 +1,7 @@
 source 'https://github.com/CocoaPods/Specs.git'
 
+platform :ios, '13.0'
+
 use_frameworks!
 
 def shared_pods
